Implement Kc and Threshold velocity for triggering the dithering torque
The threshold velocity should be the one relative to the slip/stick transition of the gearbox, i.e. the velocity under which we are in the compression phase (Dahl model) of the transmission, i.e. compression of the spring between the gearbox input (motor pos) and the gearbox output (joint pos). We choose the sliding phase as a velocity detectable on the joint side = q_motor / G ~ 0.1 deg/s.
For this reason, the input tuneable vel threshold shall be expressed in 1/100 degrees / s.

- use `config->scale` input as 1/100 degrees/s, convert to fullscaleticks/s on joint side and set
  `o->ditheringMotorVel`.
- `o->Kff` is the factor such that dither = Kff*Stiction
- Reassigned `config->offset` to the Coulomb friction o->Kc (was previously the PID "offset" K0).
- Kc input is expressed in Newton.Meters and converted to match the format of `trq_ref`.
- Refactor `PID_do_friction_comp(PID(...)` in order to integrate Kc and split the 2 states: dither or max(stiction,coulomb-viscous).
(This solution does not give good results: for low ditheringMotorVel=10, whatever the value of the dither factor; for high ditheringMotorVel=100,
 dither is the only compensation within ±1deg/s velocity, and it is harder to start the motion than without dithering.)